🏷️ 型付け
